requirements.txt
matplotlib==3.2.2
numpy==1.19.2
opencv_contrib_python==4.5.4.58
opencv_python==4.4.0.44
pandas==1.0.5
Pillow==9.5.0
scipy==1.5.0
torch==2.0.1
yolov5==7.0.12
